Add unit tests for trusty-interface and refactor.
ProxyState is promoted to its own header, and MessageProcessor has its
low-level socket code extracted into a new UnixSocket class.  This will
let us test the fiddly bits so that they retain their nascent stability.
