+++ trusty-test/ProxyStateTest.h
@@ -1,0 +1,36 @@
+// Amanuensis - Web Traffic Inspector
+//
+// Copyright (C) 2017 Benjamin Bader
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+#ifndef PROXYSTATETEST_H
+#define PROXYSTATETEST_H
+
+#include <QObject>
+
+class ProxyStateTest : public QObject
+{
+    Q_OBJECT
+
+public:
+    ProxyStateTest();
+
+private Q_SLOTS:
+    void serialize();
+
+    void deserialize_empty_host();
+};
+
+#endif // PROXYSTATETEST_H
